Create directory for notification_email and put template_ids in their own YAML file
