+++ modules/simple_forms_api/app/services/simple_forms_api/notification_email/template_ids.yml
@@ -1,0 +1,46 @@
+vba_21_0845:
+  confirmation: form21_0845_confirmation_email_template_id
+  error: form21_0845_error_email_template_id
+  received: form21_0845_received_email_template_id
+vba_21p_0847:
+  confirmation: form21p_0847_confirmation_email_template_id
+  error: form21p_0847_error_email_template_id
+  received: form21p_0847_received_email_template_id
+vba_21_0966:
+  confirmation: form21_0966_confirmation_email_template_id
+  error: form21_0966_error_email_template_id
+  received: form21_0966_received_email_template_id
+vba_21_0966_intent_api:
+  received: form21_0966_itf_api_received_email_template_id
+vba_21_0972:
+  confirmation: form21_0972_confirmation_email_template_id
+  error: form21_0972_error_email_template_id
+  received: form21_0972_received_email_template_id
+vba_21_4142:
+  confirmation: form21_4142_confirmation_email_template_id
+  error: form21_4142_error_email_template_id
+  received: form21_4142_received_email_template_id
+vba_21_10210:
+  confirmation: form21_10210_confirmation_email_template_id
+  error: form21_10210_error_email_template_id
+  received: form21_10210_received_email_template_id
+vba_20_10206:
+  confirmation: form20_10206_confirmation_email_template_id
+  error: form20_10206_error_email_template_id
+  received: form20_10206_received_email_template_id
+vba_20_10207:
+  confirmation: form20_10207_confirmation_email_template_id
+  error: form20_10207_error_email_template_id
+  received: form20_10207_received_email_template_id
+vba_40_0247:
+  confirmation: form40_0247_confirmation_email_template_id
+  error: form40_0247_error_email_template_id
+  received:
+vba_40_10007:
+  confirmation:
+  error: form40_10007_error_email_template_id
+  received:
+vba_26_4555:
+  confirmation: form26_4555_confirmation_email_template_id
+  rejected: form26_4555_rejected_email_template_id
+  duplicate: form26_4555_duplicate_email_template_id
